third_party/WebKit/LayoutTests/fast/forms/select/select-value-null.html

<!DOCTYPE html>
<title>HTMLSelectElement.value null handling</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
</script>
<select>
<option id="option1" value=""></option>
<option id="option2" value="null"></option>
</select>
<div id="log"></div>
<script>
test(function()
{
    var select = document.querySelector('select');
    assert_equals(select.value, "");
    assert_true(document.getElementById("option1").selected);
    select.value = null;
    assert_equals(select.value, "null");
    assert_true(document.getElementById("option2").selected);
}, "Setting HTMLSelectElement.value to null");
</script>
